[arcane,materials] Déplace la définition de 'ConstituentItemBase' et 'ComponentItemSharedInfo' dans leur propre fichier.
