+++ arcane/src/arcane/core/srcs.cmake
@@ -3,6 +3,8 @@
 set(ARCANE_MATERIALS_SOURCES
   materials/CellToAllEnvCellConverter.h
   materials/ConstituentItem.h
+  materials/ConstituentItemBase.h
+  materials/ConstituentItemSharedInfo.h
   materials/ConstituentItemLocalId.h
   materials/ConstituentItemLocalId.cc
   materials/MaterialsCoreGlobal.h
